draft: draft version of poseidon2 hash
