+++ Cargo.toml
@@ -14,6 +14,9 @@ rust-version = "1.80.0"
 [dependencies]
 sha2 = "0.10"
 
+# poseidon hash
+zkhash = { git = "https://github.com/HorizenLabs/poseidon2.git", rev = "bb476b9" }
+
 [dev-dependencies]
 rustc-hex = "2"
