+++ src/lib.rs
@@ -1,98 +1,2 @@
-use sha2::{Digest, Sha256};
-use std::sync::LazyLock;
-
-/// Length of a SHA256 hash in bytes.
-pub const HASH_LEN: usize = 32;
-
-/// Returns the digest of `input`.
-pub fn hash(input: &[u8]) -> Vec<u8> {
-    let mut hasher = Sha256::new();
-    hasher.update(input);
-    hasher.finalize().to_vec()
-}
-
-/// Hash function returning a fixed-size array (to save on allocations).
-pub fn hash_fixed(input: &[u8]) -> [u8; HASH_LEN] {
-    let mut hasher = Sha256::new();
-    hasher.update(input);
-    hasher.finalize().into()
-}
-
-/// Compute the hash of two slices concatenated.
-pub fn hash32_concat(h1: &[u8], h2: &[u8]) -> [u8; 32] {
-    let mut hasher = Sha256::new();
-    hasher.update(h1);
-    hasher.update(h2);
-    hasher.finalize().into()
-}
-
-pub trait Sha256Context {
-    fn new() -> Self;
-
-    fn update(&mut self, bytes: &[u8]);
-
-    fn finalize(self) -> [u8; HASH_LEN];
-}
-
-/// Context for incremental hashing
-pub struct Context(Sha256);
-
-impl Context {
-    pub fn new() -> Self {
-        Self(Sha256::new())
-    }
-
-    pub fn update(&mut self, bytes: &[u8]) {
-        self.0.update(bytes);
-    }
-
-    pub fn finalize(self) -> [u8; HASH_LEN] {
-        self.0.finalize().into()
-    }
-}
-
-/// The max index that can be used with `ZERO_HASHES`.
-#[cfg(feature = "zero_hash_cache")]
-pub const ZERO_HASHES_MAX_INDEX: usize = 48;
-
-#[cfg(feature = "zero_hash_cache")]
-/// Cached zero hashes where `ZERO_HASHES[i]` is the hash of a Merkle tree with 2^i zero leaves.
-pub static ZERO_HASHES: LazyLock<Vec<[u8; HASH_LEN]>> = LazyLock::new(|| {
-    let mut hashes = vec![[0; HASH_LEN]; ZERO_HASHES_MAX_INDEX + 1];
-
-    for i in 0..ZERO_HASHES_MAX_INDEX {
-        hashes[i + 1] = hash32_concat(&hashes[i], &hashes[i]);
-    }
-
-    hashes
-});
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rustc_hex::FromHex;
-
-    #[cfg(target_arch = "wasm32")]
-    use wasm_bindgen_test::*;
-
-    #[cfg_attr(not(target_arch = "wasm32"), test)]
-    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
-    fn test_hashing() {
-        let input: Vec<u8> = b"hello world".as_ref().into();
-
-        let output = hash(input.as_ref());
-        let expected_hex = "b94d27b9934d3e08a52e52d7da7dabfac484efe37a5380ee9088f7ace2efcde9";
-        let expected: Vec<u8> = expected_hex.from_hex().unwrap();
-        assert_eq!(expected, output);
-    }
-
-    #[cfg(feature = "zero_hash_cache")]
-    mod zero_hash {
-        use super::*;
-
-        #[test]
-        fn zero_hash_zero() {
-            assert_eq!(ZERO_HASHES[0], [0; 32]);
-        }
-    }
-}
+pub mod poseidon2;
+pub mod sha2;
